FIX: Allow configuration of IPMI privilege level
This change adds a priv_level argument to the hwmgmt specification.
The value of this parameter will be used as the requested privilege
level on IPMI requests.  Without this parameter, IPMI requests will
fail if the hwmgmt.user does not have ADMINISTRATOR privilege.  To
preserve the original semantics, the default value for the parameter
is ADMINISTRATOR.

signed-off-by: dave kormann <davek@research.att.com>
Change-Id: I8adf8d32510801914f4e8438190c52f5e9524209