Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > how would this work after in-sourcing the same plugin?
> > the whole idea of setting different default value from a REC CM plugin was so that we don't need to this backward incompatible change in the shared code
> > 
> > what happened with the REC CM plugin approach?
> 
> It was not working as it was imagined.
> It seems this `cmdatahandlers/caas/config.py` runs earlier than `recuserconfighandlers/reccaashandler/reccaashandler.py`. If both code pieces initialize a variable if it is not set earlier, then the `cmdatahandlers/caas/config.py` wins.
> 
> Currently this config.py has full of REC specific values. It seems this concept was never implemented/kept.

this was before we figured out the elegant way, which is the handling data through CM plugins rather than hard-coding. it doesn't mean this one need not be reworked when an insource happens, it only means we haven't done it yet.
let's not add to the rework if it is not needed. the solution is removing this variable from the shared code IMO
then REC sets it from its own handler, and NCIR also sets it from its own handler. that way there never will be clashes

Patch-set: 1